requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.4
djangorestframework==3.13.1
numpy==1.22.3
opencv-contrib-python-headless==4.5.5.64
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
